pyproject.toml
[project]
name = "image-to-excel-mcp"
version = "0.1.0"
description = "Convert table images to Excel files using MCP protocol and Qwen-VL model."
authors = [
    { name = "workQ", email = "qihao967@gmail.com" }
]
dependencies = [
    "fastmcp",
    "openai>=1.0.0",
    "Pillow",
    "pandas",
    "python-dotenv",
    "uvicorn",
    "starlette",
    "anyio"
]
requires-python = ">=3.9,<3.13"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools]
include-package-data = true
